split host into separate file
